The functions in primary.R were updated so that they use the apropriate input data:
Each experimental condition is treated as its own entity and so scaling the data together is inapropriate as it forces the two conditions to appear more similar than they actually are. BICtune, stability selection, and getNetworks were updated so that they pull the data that was scaled independently and stored in the 'scaled_expression_data' element of the assays slot. This was already the case for BICtune, getNetworks, and stability selection without subsampling. However, the assays slot was modified so that data scaled together is no longer stored in the assays slot, so the changes here also reflect that. Additionally, runNetGSA now used the log transformed input data instead of the raw peak intensity/concentrations. The old, commented out code has also been removed.