updated createDNEAobject() to take as input scaled data:
restructure_scaled_input_data() function was created to manipulate scaled data to initiate DNEAobj when only scaled data is supplied. createDNEAobject() now checks for expression_data parameter and operates normally if present. If scaled_expression_data also provided, it is input to the assays slot under 'scaled_expression_data' and the log-scaled data is moved to 'DNEA_scaled_data'. If only scaled_expression_data is provided, restructure_scaled_input_data is used to manipulate data for initiation of DNEAobj and a warning is thrown to the user recommending that they input raw peak intensity/concentration data for more accurate results.